Added 3 More TestMethods for SandBox.
