+++ transaction_integration_test.go
@@ -457,48 +457,6 @@ func TestTransactionCreateFromPaymentMethodCode(t *testing.T) {
 	}
 }
 
-func TestSettleTransaction(t *testing.T) {
-	t.Parallel()
-
-	old_environment := testGateway.Environment
-
-	txn, err := testGateway.Transaction().Create(&Transaction{
-		Type:   "sale",
-		Amount: randomAmount(),
-		CreditCard: &CreditCard{
-			Number:         testCreditCards["visa"].Number,
-			ExpirationDate: "05/14",
-		},
-	})
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	txn, err = testGateway.Transaction().SubmitForSettlement(txn.Id, txn.Amount)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	testGateway.Environment = Production
-
-	_, err = testGateway.Transaction().Settle(txn.Id)
-	if err.Error() != "Operation not allowed in production environment" {
-		t.Log(testGateway.Environment)
-		t.Fatal(err)
-	}
-
-	testGateway.Environment = old_environment
-
-	txn, err = testGateway.Transaction().Settle(txn.Id)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	if txn.Status != "settled" {
-		t.Fatal(txn.Status)
-	}
-}
-
 func TestTrxPaymentMethodNonce(t *testing.T) {
 	t.Parallel()
 
@@ -538,7 +496,7 @@ func TestTransactionCreateSettleAndFullRefund(t *testing.T) {
 		t.Fatal(err)
 	}
 
-	txn, err = testGateway.Transaction().Settle(txn.Id)
+	txn, err = testGateway.Testing().Settle(txn.Id)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -559,7 +517,7 @@ func TestTransactionCreateSettleAndFullRefund(t *testing.T) {
 		t.Fatal(x)
 	}
 
-	refundTxn, err = testGateway.Transaction().Settle(refundTxn.Id)
+	refundTxn, err = testGateway.Testing().Settle(refundTxn.Id)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -614,7 +572,7 @@ func TestTransactionCreateSettleAndPartialRefund(t *testing.T) {
 		t.Fatal(err)
 	}
 
-	txn, err = testGateway.Transaction().Settle(txn.Id)
+	txn, err = testGateway.Testing().Settle(txn.Id)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -635,7 +593,7 @@ func TestTransactionCreateSettleAndPartialRefund(t *testing.T) {
 		t.Fatal(x)
 	}
 
-	refundTxn, err = testGateway.Transaction().Settle(refundTxn.Id)
+	refundTxn, err = testGateway.Testing().Settle(refundTxn.Id)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -651,46 +609,6 @@ func TestTransactionCreateSettleAndPartialRefund(t *testing.T) {
 
 	if err.Error() != "Refund amount is too large." {
 		t.Fatal(err)
-	}
-}
-
-func TestTransactionCreateSettleCheckCreditCardDetails(t *testing.T) {
-	t.Parallel()
-
-	amount := NewDecimal(10000, 2)
-	txn, err := testGateway.Transaction().Create(&Transaction{
-		Type:   "sale",
-		Amount: amount,
-		CreditCard: &CreditCard{
-			Number:         testCreditCards["discover"].Number,
-			ExpirationDate: "05/14",
-		},
-	})
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	if txn.PaymentInstrumentType != "credit_card" {
-		t.Fatalf("Returned payment instrument doesn't match input, expected %q, got %q",
-			"credit_card", txn.PaymentInstrumentType)
-	}
-	if txn.CreditCard.CardType != "Discover" {
-		t.Fatalf("Returned credit card detail doesn't match input, expected %q, got %q",
-			"Visa", txn.CreditCard.CardType)
-	}
-
-	txn, err = testGateway.Transaction().SubmitForSettlement(txn.Id, txn.Amount)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	txn, err = testGateway.Transaction().Settle(txn.Id)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	if txn.Status != "settled" {
-		t.Fatal(txn.Status)
 	}
 }
